==> Project/0 - Import and pickle the dataset.ipynb <==
# Extraction of the UCDP Dataset

import os

import numpy as np
import pandas as pd
import geopandas as gpd
import pickle
from tqdm import tqdm
from bs4 import BeautifulSoup

from IPython.display import display

CONFLICT_DATA_PATH = os.path.join("data", "ged171.csv")
raw_conflict_df = pd.read_csv(CONFLICT_DATA_PATH)
raw_conflict_df.head(5)

raw_conflict_df.columns

KEPT_COLUMNS= ['id', 'year', 'type_of_violence', 'conflict_new_id',
       'conflict_name', 'side_a_new_id', 'gwnoa',
       'side_a', 'gwnob', 'side_b_new_id','side_b',
       'where_prec', 'where_coordinates', 'adm_1', 'adm_2',
       'latitude', 'longitude', 'geom_wkt','country',
       'country_id', 'region','date_start', 'date_end','deaths_a', 'deaths_b',
       'deaths_civilians',
       'deaths_unknown', 'best'] 

conflict_df = raw_conflict_df[KEPT_COLUMNS]

conflict_df = conflict_df.replace(to_replace="-[0-9]{2}-[0-9]{2}", value="", regex=True)

conflict_df.head(5)

with open(os.path.join('pickle','conflict.pickle'), 'wb') as out:
    pickle.dump(conflict_df, out)

with open(os.path.join('pickle', 'conflict.pickle'), 'rb') as data_source:
    conflict_df = pickle.load(data_source)
