==> Project/2 - Preparing UNHCR, GDP and HDI datasets.ipynb <==
import os

import numpy as np
import pandas as pd
import pickle
from tqdm import tqdm

import re
import unidecode

from IPython.display import display

# Prepare the Refugee dateset

## Extraction and cleanup of the refugee data

### Import the data, remove unused fields and reorganize

First we import the dataset downloaded from [the UNHCR website](http://www.unhcr.org/) making sure the different fields have the appropriate type:


REFUGEE_DATA_PATH = os.path.join("data", "unhcr_refugee.csv")
RAW_COLUMN_NAMES = ["year", "country_dest", "origin", "refugee",
                "asylum", "returned_refugee", "internally_displaced", "returned_idp",
                "stateless", "others", "total"
               ]
RAW_COLUMN_TYPE = {"year": int, "coutry_dest" : object, "origin" : object, "refugee" : float,
               "asylum" : float, "returned_refugee" : float, "idp" : float, "returned_idp" : float,
               "stateless" : float, "others" : float, "total" : float
              }
raw_refugee_df = pd.read_csv(REFUGEE_DATA_PATH,
                             skiprows=4,
                             names=RAW_COLUMN_NAMES,
                             dtype=RAW_COLUMN_TYPE,
                             na_values=["*"])
raw_refugee_df.head(5)

We replace the missing values with 0

raw_refugee_df.fillna(value=0, inplace=True)
raw_refugee_df.head(5)

### Group by year and origin of migration

We drop all the origins that are `Various/Unknown`: we are interested in the countries of origin of the migrants, so this identifient is useless to our analysis.
We also drop the returned columns because it is symptomatic of past refugee and doesn't really fit in our analysis

raw_refugee_df = raw_refugee_df[(raw_refugee_df.origin != 'Various/Unknown') & (raw_refugee_df.origin != 'Stateless')]

try:
    raw_refugee_df.drop(['returned_refugee', 'returned_idp'], axis=1, inplace=True)
except: # avoid error if we re-run this code
    pass

raw_refugee_df.head(5)

We reorganize the data to have values grouped by year and country of origin. We are only interested in the country of origin as we want to study the correlation between the conflict and the outgoing flows of population

REFUGEE_COLUMNS = ["year", "origin", "refugee", "asylum", "internally_displaced", "stateless", "others", "total"]
refugee_df = pd.DataFrame(columns=REFUGEE_COLUMNS)

for year in tqdm(raw_refugee_df.year.unique()):
    for origin in raw_refugee_df[raw_refugee_df.year == year].origin.unique():
        index = (raw_refugee_df.year == year) & (raw_refugee_df.origin == origin)
        temp_df_no_dest = raw_refugee_df[index].drop(["country_dest"], axis=1)
        sum_series = temp_df_no_dest.sum(numeric_only=True)
        # drop the row if the column of interest are zero 
        # we don't check total, because it might take into account returne, which we dropped
        if (sum_series[1:5] == 0).all():
            continue
            
        sum_series["year"] = year
        sum_series["origin"] = origin
        sum_series["total"] = sum_series[1:5].sum()
        
        refugee_df = refugee_df.append(sum_series, ignore_index=True)
        
refugee_df.head(5)

## Merge the Refugee Dataset with the Conflicts dataset

The GWNO in the conflict dataset corresponds to the the *Gleditsch and Ward* number which identifies uniquely every independent state. We download two datafiles associating this number of the country name. Here we import them, clean them and store them into a dataframe to be able to merge later:

def extract_gnwo_countries_to_df():
    """Extract the countries from the gnwo and their id from the gnwo files"""
    countries_list = []
    # First gnwo file, contains id, code, name for all gnow numbers
    with open(os.path.join("data", "gnwo.txt"), "r") as gnow:
        for line in gnow:
            split_line = re.split(r'\t+', line)
            countries_list.append(split_line[0:3])
            
    # We add the gnwo2 file, it contains more countries
    with open(os.path.join("data", "gnwo2.txt"), "r") as gnow2:
        for line in gnow2:
            val = line.strip().split(";")
            split_line = [val[0], None, val[1]]
            countries_list.append(split_line)
        
    return pd.DataFrame(countries_list, columns=["id", "code" ,"name"])
countries = extract_gnwo_countries_to_df()
display(countries.head(2))
display(countries.tail(2))

def get_country_id(country_df, country, use_special_case_dict=True):
    """Check if the country name is in the standard countries dataset
       Particular conditions must be respected in the balkans in the years of interest.
       Returns the gwn corresponding to the provided country name
    """
    country_id=None
    # Force ascii names of country
    country = unidecode.unidecode(country)
    
    # Strip leading/trailing whitespaces
    country = country.strip()
    
    # Manage diminutive problem
    country = re.sub(r"Dem\.", "Democratic", country)
    country = re.sub(r"Rep\.", "Republic", country)
    
    # Manage China
    country = re.sub(r".*China.*", "China", country)
    
    # Manage special cases
    # Note: We decided to do this instead of dropping the values
    #       The values will be reported to their "controlling" states
    special_case_dict = {
        "Western Sahara" : 600, # Contested, but set as Morroco
        "Bolivia (Plurinational State of)" : 145, # Bolivia
        "Palestinian" : 666, # Israel
        "Serbia and Kosovo (S/RES/1244 (1999))" :345, # Serbia 
        "Tibetan" : 711, # Tibet
        "Cabo Verde" : 402, # Cape Verde
        "French Guiana" : 220, # France
        "The former Yugoslav Republic of Macedonia" : 343, # Macedonia
        "Puerto Rico" : 2, # USA
        "Martinique" : 220, # France
        "Bermuda" : 200, # United Kingdom
        "Micronesia (Federated States of)" : 987, # Micronesia
        "Cayman Islands" : 200, # United Kingdom
        "Gibraltar" : 200, # United Kingdom
        "Turks and Caicos Islands" : 200, # United Kingdom
        "Niue" : 920, # New Zealand
        "French Polynesia" : 220, # France
        "Holy See (the)" : None, # Vatican
        "New Caledonia" : 220, # France
        "Cook Islands" : 920, # New Zealand
        "Curacao" : 210, # Netherlands
        "British Virgin Islands" : 200, # United Kingdom
        "Guadeloupe" : 220, # France
        "Norfolk Island" : 900, # Australia
        "Wallis and Futuna Islands " : 220, # France
        "Saint-Pierre-et-Miquelon" : 220, # France
        "Svalbard and Jan Mayen" : 385, # Norway
    }
    
    check_result = country_df.name.str.contains(country, regex=False) 
    index = country_df.index[check_result]
    if index.empty:
        if use_special_case_dict:
            try:
                country_id = special_case_dict[country]
            except KeyError:
                print("Warning: failed to find index of country {}".format(country))
                country_id = None
        else:
            print("Warning: failed to find index of country {}".format(country))
    else:
        country_id = country_df.loc[index[0], "id"]
        
    return country_id

We create a Series containing the gwn of each country in the refugee dataset:

id_with_index_list = []
for row in tqdm(refugee_df.itertuples()):
    country_id = get_country_id(countries, row.origin)
    index = row.Index
    id_with_index_list.append([index, country_id])
country_id_df = pd.DataFrame(id_with_index_list, columns=["index", "country_id"])
country_ids = pd.Series(country_id_df.country_id, index=country_id_df.index)
display(country_ids.head(5))

if not "country_id" in refugee_df.columns:
    refugee_df = pd.concat([refugee_df, country_ids], axis=1)
try:
    refugee_df = refugee_df.drop("origin", axis=1)
except:
    pass

refugee_df["country_id"] = pd.to_numeric(refugee_df["country_id"])
display(refugee_df.head(5))

We want to merge rows with duplicated `year` and `country_id`, so we have one value per tuple `(country, year)`

# Identify the duplicates
year_id_dup = refugee_df.duplicated(subset=["year", "country_id"])
not_dup_refugee_df = refugee_df.loc[~year_id_dup]
dup_refugee_df = refugee_df.loc[year_id_dup]

refugee_df = dup_refugee_df.copy()

for row in tqdm(dup_refugee_df.itertuples()):
    values = row[1:]
    not_dup_loc = (not_dup_refugee_df["year"] == row.year) & (not_dup_refugee_df["country_id"] == row.country_id)
    # Add values
    not_dup_refugee_df.loc[not_dup_loc] = not_dup_refugee_df.loc[not_dup_loc].add(values)
    # Year and country_id where added, set the proper values
    not_dup_refugee_df.loc[not_dup_loc, "year"] = row.year
    not_dup_refugee_df.loc[not_dup_loc, "country_id"] = row.country_id

not_dup_refugee_df.duplicated(subset=["year", "country_id"]).any()

refugee_df = not_dup_refugee_df

with open(os.path.join('pickle', 'refugee.pickle'), 'wb') as out:
    pickle.dump(refugee_df, out)

with open(os.path.join('pickle', 'refugee.pickle'), 'rb') as data_source:
    refugee_df = pickle.load(data_source)

# Prepare the GDP Dataset 

We want to combine the gross domestic product with the conflict dataset to analyze whether there is a correlation between the economic situation of a country and its involvement in armed conflicts. We obtained the GDP data from [UNdata](http://data.un.org)

gdp_df = pd.read_csv(os.path.join("data","UNdata_gdp.csv"), usecols=[0,1,3])
gdp_df[["country", "year", "gdp"]] = gdp_df[["Country or Area", "Year", "Value"]]
gdp_df = gdp_df.drop(["Country or Area", "Year", "Value"], axis=1)

def get_country_id_gdp(country_df, country, not_indexed=[]):
    """Check if the country name is in the standard countries dataset
       Particular conditions must be respected in the balkans in the years of interest
    """
    country_id=None
    # Force ascii names of country
    country = unidecode.unidecode(country)
    
    # Strip leading/trailing whitespaces
    country = country.strip()
    
    # Manage diminutive problem
    country = re.sub(r"Dem\.", "Democratic", country)
    country = re.sub(r"Rep\.", "Republic", country)
    
    # Manage China
    country = re.sub(r".*China.*", "China", country)

    check_result = country_df.name.str.contains(country, regex=False) 
    index = country_df.index[check_result]
    if index.empty:
        print("Warning: failed to find index of country '{}'".format(country))
        # We do no management of None values on gdp
        country_id = None
    else:
        country_id = country_df.loc[index[0], "id"]
        
    if not_indexed:    
        return country_id, not_indexed
    else:
        return country_id
        

display(gdp_df.head(5))

country_id = []
for row in gdp_df.itertuples():
    try:
        country_id.append(get_country_id(countries, row.country, use_special_case_dict=False))
    except:
        country_id.append(None)
country_id = pd.Series(country_id, name="country_id")

if not "country_id" in gdp_df.columns:
    gdp_df = pd.concat([gdp_df, country_id], axis=1)
    gdp_df = gdp_df.drop("country", axis=1)
gdp_df.dropna(inplace=True)
display(gdp_df.head(5))

gdp_df["year"] = pd.to_numeric(gdp_df["year"])
gdp_df["country_id"] = pd.to_numeric(gdp_df["country_id"])
gdp_df = gdp_df.dropna(axis=0, how="any")
display(gdp_df.head(5))

# Drop duplicated gdp
gdp_df.drop_duplicates(["year", "country_id"], inplace=True)
print(gdp_df.duplicated(["year", "country_id"]).sum())

gdp_df = gdp_df.set_index(["year", "country_id"]).sort_index()

with open(os.path.join('pickle','gdp.pickle'), 'wb') as out:
    pickle.dump(gdp_df, out)

# Prepare the HDI Dataset

In addition to the GDP of country, we would like to use the [Human Development Index](http://hdr.undp.org/en/content/human-development-index-hdi) to perform a similar analysis.

hdi_df = pd.read_csv(os.path.join("data","hdi_undp.csv"), skiprows=[0])

display(hdi_df.head(5))

countries = extract_gnwo_countries_to_df()
country_id = []
for row in hdi_df.itertuples():
    local_country_id = get_country_id(countries, row.Country, False)
    country_id.append(local_country_id)
hdi_country_id = pd.Series(country_id, name="country_id")

if not "country_id" in hdi_df.columns:
    hdi_df = pd.concat([hdi_df, hdi_country_id], axis=1)
display(hdi_df.head(1))

if "Country" in hdi_df.columns:
    hdi_df = hdi_df.drop(["HDI Rank", "Country"], axis=1)
display(hdi_df.head(1))

col_dict_mapper= {}
for col in hdi_df.columns:
    try:
        int(col)
    except:
        pass
    else:
        col_dict_mapper[col] = "y{}".format(col)
hdi_df = hdi_df.rename(index=str, columns=col_dict_mapper)
display(hdi_df.head(1))

new_hdi_df = pd.DataFrame([], columns=["year", "hdi", "country_id"])
for row in tqdm(hdi_df.itertuples()):
    for year in range(1990,2016):
        local_series = pd.Series([year, row[year-1989], row.country_id],
                                 index=["year", "hdi", "country_id"])
        new_hdi_df = new_hdi_df.append(local_series, ignore_index=True)
display(new_hdi_df.head(5))

new_hdi_df = new_hdi_df.dropna()

# Make sure everything is numeric
new_hdi_df.year =  pd.to_numeric(new_hdi_df.year)
new_hdi_df.hdi =  pd.to_numeric(new_hdi_df.hdi)
new_hdi_df.country_id =  pd.to_numeric(new_hdi_df.country_id)

new_hdi_df.head(5)

with open(os.path.join('pickle','hdi.pickle'), 'wb') as out:
    pickle.dump(new_hdi_df, out)
