==> Project/3 - Merge all the datasets.ipynb <==
# Combine the datasets in a single dataframe

import os

import numpy as np
import pandas as pd
import networkx as nx
import pickle
from tqdm import tqdm
import geopandas as gpd

from scipy import stats

import re
import unidecode

import matplotlib.pyplot as plt
%matplotlib inline

## Merge with UNHCR dataset

with open(os.path.join("pickle", 'conflict.pickle'), 'rb') as data_source:
    conflict_df = pickle.load(data_source)

with open(os.path.join("pickle", 'refugee.pickle'), 'rb') as data_source:
    refugee_df = pickle.load(data_source)

The goal of this section is to create a dataframe that links the displacement information given by the UNHCR with the events and deaths given by the GED dataset. This will be grouped by year and will keep track of the number of events and of the event identifiers.

The columns will be:
 - `country_id`
 - `year` 
 - `number_of_events` the number of events in a country in a year
 - `events_id` an array containing the id of each event accounted for
 - `displacement_extern` the sum of `[refugee, asylum, stateless, others]`
 - `displacement_intern` the value of internally_displaced
 - `total_displacement` 
 - `deaths`

COUNTRY_HUMAN_COST_DF_COLUMNS = [
    "country_id",
    "year",
    "number_of_events",
    "events_id",
    "displacement_extern",
    "displacement_intern",
    "total_displacement",
    "deaths"
]

### Manage index and data types

We sort and modify the indexing of our dataframes for easier and more performant data wrangling

# Sort conflict dataframe so we can easily iter through it
sorted_conflict_df = conflict_df.sort_values(["year", "country_id"], axis=0)
sorted_conflict_df.head(2)

refugee_df["year"] = pd.to_numeric(refugee_df["year"])
refugee_df["country_id"] = pd.to_numeric(refugee_df["country_id"])

# Index refugee_df for easier search
try:
    refugee_df = refugee_df.set_index(["year", "country_id"])
except: # if we run this two times
    pass
# Get the displacement_extern value
refugee_df["displacement_extern"] = refugee_df.apply(\
                                    lambda row: row.refugee + row.asylum + row.stateless + row.others, axis=1)
refugee_df.head(1)

refugee_df = refugee_df.sort_index()
refugee_df.head(5)

def get_total_deaths_from_event_ids(conflict_df, event_ids):
    """This function will return the total deaths associated with a list of event id"""
    deaths = 0
    for event in event_ids:
        index = conflict_df.index[conflict_df.id == event].tolist()[0]
        deaths += conflict_df.loc[index, "best"]
    return deaths

def get_append_series(conflict_df, refugee_df, year, country_id, event_ids=[]):
    """This function will create a serie to be appended to the country_human_cost_df with the proper values
       at each column
    """
    try:
        displacement_extern = refugee_df.displacement_extern[year, country_id]
        displacement_intern = refugee_df.internally_displaced[year, country_id]
        total_displacement  = refugee_df.total[year, country_id]
    except KeyError: # No value for this key
        displacement_extern = 0
        displacement_intern = 0
        total_displacement  = 0
        
    deaths = get_total_deaths_from_event_ids(conflict_df, event_ids)
    
    data = [ country_id, year, len(event_ids), event_ids,
             displacement_extern, displacement_intern,
             total_displacement, deaths
           ] 
    series_to_append = pd.Series(data = data, index=COUNTRY_HUMAN_COST_DF_COLUMNS)
    return series_to_append

We create the `country_human_cost_df` and fill it country by country, with different values for each year

# Init variables
current_year = None
current_country_id = None
event_ids = []

country_human_cost_df = pd.DataFrame(columns=COUNTRY_HUMAN_COST_DF_COLUMNS) 
for row in tqdm(sorted_conflict_df.itertuples()):
    # Check if we are still in the same year and country
    if (current_year != row.year) or (current_country_id != row.country_id):
        # If we registered event, add a row to the country_human_cost_df
        if event_ids:
            series_to_append = get_append_series(conflict_df, refugee_df, current_year, current_country_id, event_ids)
            country_human_cost_df = country_human_cost_df.append(series_to_append, ignore_index=True)
        
        # Set the variables to the desired values
        event_ids = []
        current_year = row.year
        current_country_id = row.country_id
        
    # Add the event to the list
    event_ids.append(row.id)

country_human_cost_df["country_id"] = pd.to_numeric(country_human_cost_df["country_id"])
country_human_cost_df["year"] = pd.to_numeric(country_human_cost_df["year"])
country_human_cost_df.head(5)

We change the index again after merging:

try:
    country_human_cost_df = country_human_cost_df.set_index(["year", "country_id"])
except:  # In case cell is re-run
    pass
country_human_cost_df = country_human_cost_df.sort_index()
country_human_cost_df.head(5)

We also take into account migrants displacement for countries where there was no conflict

new_country_human_cost_df = country_human_cost_df
for row in tqdm(refugee_df.itertuples()):
    try:
        country_human_cost_df.loc[row.Index]
    except KeyError: # The row is not in the country_human_cost_df
        series_to_append = get_append_series(conflict_df, refugee_df, row.Index[0], row.Index[1]) 
        df_to_append = pd.DataFrame([series_to_append.transpose()])
        df_to_append = df_to_append.set_index(["year", "country_id"])
        new_country_human_cost_df = new_country_human_cost_df.append(df_to_append)
    except TypeError:
        continue


country_human_cost_df = new_country_human_cost_df

country_human_cost_df = country_human_cost_df.sort_index()
country_human_cost_df.head(5)

## Merge with GDP dataset

with open(os.path.join("pickle", 'gdp.pickle'), 'rb') as data_source:
    gdp_df = pickle.load(data_source)

gdp_df.head(5)

try: # Set index if not already done
    gdp_df = gdp_df.set_index(["year", "country_id"])
except:
    pass

gdp_df.index.duplicated().sum()

human_cost_gdp_df = pd.merge(country_human_cost_df, gdp_df, how='outer',
                             left_index=True, right_index=True, sort=True)

human_cost_gdp_df.fillna(value=0, inplace=True)
    
human_cost_gdp_df.head(5)

country_human_cost_df = human_cost_gdp_df

## Merge with the HDI dataset

with open(os.path.join("pickle", 'hdi.pickle'), 'rb') as data_source:
    hdi_df = pickle.load(data_source)

hdi_df.head(2)

hdi_df = hdi_df.set_index(["year", "country_id"])
print(len(hdi_df))

hdi_duplicated_index = hdi_df.index.duplicated()
hdi_df = hdi_df.loc[~hdi_duplicated_index]
print(len(hdi_df))
hdi_df.head(5)

human_cost_hdi_df = pd.merge(country_human_cost_df, hdi_df, how='outer',
                             left_index=True, right_index=True, sort=True)

human_cost_hdi_df.fillna(value=0, inplace=True)
    
human_cost_hdi_df.head(5)

country_human_cost_df = human_cost_hdi_df

## Add country codes to the dataframe

We add the country codes to be able to display data on map using geopandas and other libraries

country_codes_df = pd.read_csv(os.path.join("data", "countrycode.csv"))
country_codes_df.head(5)

def extract_iso3_from_gnow(gnwo):
    """Get the iso3 country code with the gnwo value, return None if no match was found"""
    try:
        iso3 = country_codes_df.loc[country_codes_df.cown == gnwo]["wb_api3c"].values[0]
    except IndexError as e:
        # 99 = Great Columbia
        # 711 = Tibet
        # 972 = Tonga
        # 973 = Tuvalu
        if not (gnwo == 99 or gnwo == 711 or gnwo == 971 or gnwo == 972 or gnwo == 973 or pd.isnull(gnwo) or gnwo == 340):
            print(e, "with gnwo: {}".format(gnwo))
        return None
    return iso3
print(extract_iso3_from_gnow(2))

def extract_gnwo_countries_to_df():
    """Extract the countries from the gnwo and their id from the gnwo files"""
    countries_list = []
    # First gnwo file, contains id, code, name for all gnow numbers
    with open(os.path.join("data", "gnwo.txt"), "r") as gnow:
        for line in gnow:
            split_line = re.split(r'\t+', line)
            countries_list.append(split_line[0:3])
    countries_df = pd.DataFrame(countries_list, columns=["id", "code", "name"])
    countries_df["id"] = pd.to_numeric(countries_df["id"])
    countries_df = countries_df.drop_duplicates()
    return countries_df.set_index("id")
countries_list = extract_gnwo_countries_to_df()

countries_list.head(5)

# Add code column to the country_human_cost_df
countries_codes = []
for row in country_human_cost_df.itertuples():
    countries_codes.append([row.Index[0], row.Index[1], extract_iso3_from_gnow(row.Index[1])])
countries_codes = pd.DataFrame(countries_codes, columns=["year", "country_id","country_code"])\
                    .set_index(["year", "country_id"])

if not "country_code" in country_human_cost_df.columns:
    country_human_cost_df = pd.concat([country_human_cost_df, countries_codes], axis=1, join_axes=[country_human_cost_df.index])

country_human_cost_df = country_human_cost_df[~country_human_cost_df.country_code.isnull()]
country_human_cost_df.head(5)

country_human_cost_df.country_code.isnull().any()

# Compare the obtained country code with the one in the geopanda world dataset
world = gpd.read_file(gpd.datasets.get_path('naturalearth_lowres'))
world.head(5)

# Drop country for which we don't have a corresponding country code
country_code_to_drop = []
for country_code in country_human_cost_df.country_code.unique():
    if not world.iso_a3.str.contains(country_code).any():
        country_code_to_drop.append(country_code)

# Remove unrecognized countries from the dataframe
country_human_cost_df = country_human_cost_df[~country_human_cost_df.country_code.isin(country_code_to_drop)]
country_human_cost_df.head(5)

## Pickle the result

with open(os.path.join("pickle", 'country_human_cost.pickle'), 'wb') as out:
    pickle.dump(country_human_cost_df, out)

with open(os.path.join("pickle", 'country_human_cost.pickle'), 'rb') as data_source:
    country_human_cost_df = pickle.load(data_source)
